SID-944 Fix to correctly handle periods in file names when creating directory to extract to.

+++ Components/Extractor/src/test/java/edu/si/services/camel/extractor/ExtractorComponentTest.java
@@ -33,6 +33,7 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.test.junit4.CamelTestSupport;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
@@ -55,6 +56,7 @@ public class ExtractorComponentTest extends CamelTestSupport
 {
     private static final Logger LOG = LoggerFactory.getLogger(ExtractorComponentTest.class);
     private StringBuilder expectedExceptionMsg;
+    private static final boolean DELETE_AFTER_TEST = false;
 
     //TODO: Add more test!!! Need to make sure that Producer finds the correct new folder if there are multiple folders initially
     //currently we assume that an archive contains one or no compressed directory.
@@ -103,6 +105,12 @@ public class ExtractorComponentTest extends CamelTestSupport
         testExtractor("test-no-file-ext");
     }
 
+    @Test
+    public void testFileNameWithPeriods() throws Exception {
+        testExtractor("test_zip.with.periods_1.zip", false);
+        testExtractor("test_zip.with.periods_2.zip", false);
+    }
+
     public void testExtractor(String archive) throws Exception
     {
         this.testExtractor(archive, true);
@@ -128,6 +136,7 @@ public class ExtractorComponentTest extends CamelTestSupport
 
             assertNotNull("Results should not be null", body);
             assertTrue("Results should be a directory", body.isDirectory());
+            assertTrue("Directory name should match archive name", FilenameUtils.getBaseName(archive).equals(msg.getHeader("CamelFileNameOnly")));
             assertEquals("Parent directory should be 'TestCameraTrapData'", "TestCameraTrapData", body.getParentFile().getName());
             assertEquals("Directory should contain " + numFiles + " elements", numFiles, body.list().length);
 
@@ -217,7 +226,7 @@ public class ExtractorComponentTest extends CamelTestSupport
             {
                 from("direct:start")
                         .to("log:edu.si.ctingest?level=DEBUG&showHeaders=true")
-                        .to("extractor:extract?location=TestCameraTrapData")
+                        .to("extractor:extract?location=target/TestCameraTrapData")
                         .to("log:edu.si.ctingest?level=DEBUG&showHeaders=true")
                         .to("mock:result");
             }
@@ -235,13 +244,12 @@ public class ExtractorComponentTest extends CamelTestSupport
     @AfterClass
     public static void afterClass()
     {
-        try
-        {
-            FileUtils.deleteDirectory(new File("TestCameraTrapData"));
-        }
-        catch (IOException ex)
-        {
-            LOG.error(null, ex);
+        if (DELETE_AFTER_TEST) {
+            try {
+                FileUtils.deleteDirectory(new File("TestCameraTrapData"));
+            } catch (IOException ex) {
+                LOG.error(null, ex);
+            }
         }
     }
 }
